+++ frontend/app/players/[username]/page.tsx
@@ -1,29 +1,13 @@
-import { auth } from "@clerk/nextjs/server";
-
-import { BattleStadiumAPI } from "@/lib/api";
-
-function getApiClient(shouldAuth = true) {
-  if (shouldAuth) {
-    return BattleStadiumAPI(auth());
-  }
-
-  return BattleStadiumAPI();
-}
-
-async function getPlayer(username: string) {
-  const response = await getApiClient().Users.get(username);
-
-  return response.data;
-}
+import { getUser } from "@/app/data/actions";
 
 export async function generateMetadata({ params }: { params: { username: string } }) {
-  const player = await getPlayer(params.username);
+  const player = await getUser(params.username);
 
   return { title: player?.username ?? "Player" };
 }
 
 export default async function PlayerProfilePage({ params }: { params: { username: string } }) {
-  const player = await getPlayer(params.username);
+  const player = await getUser(params.username);
 
   return (
     <div>
