+++ frontend/app/players/page.tsx
@@ -1,18 +1,5 @@
-import { BattleStadiumAPI, components } from "@/lib/api";
 import PlayersTable from "@/app/players/players-table";
-
-type User = components["schemas"]["User"];
-
-async function fetchPlayers(): Promise<User[]> {
-  try {
-    return (await BattleStadiumAPI().Users.list()).data ?? [];
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error("Failed to fetch players:", error);
-
-    return [];
-  }
-}
+import { getUsers } from "../data/actions";
 
 const columns = [
   {
@@ -26,7 +13,7 @@ const columns = [
 ];
 
 export default async function PlayersPage() {
-  const players = await fetchPlayers();
+  const players = await getUsers();
 
   return <PlayersTable columns={columns} players={players} />;
 }
